move all copy methods to FileUtil
